rename tr to table-row for clarity
