+++ libs/designsystem/src/lib/components/data-table/index.ts
@@ -1,4 +1,4 @@
 export { TableComponent } from './table/table.component';
-export { TableRowComponent } from './tr/tr.component';
+export { TableRowComponent } from './table-row/table-row.component';
 
 export { DataTableModule } from './data-table.module';
