+++ libs/designsystem/src/lib/components/data-table/table-row/table-row.component.scss
