+++ libs/designsystem/src/lib/components/data-table/table-row/table-row.component.ts
@@ -4,7 +4,7 @@ import { Component, HostBinding, Input } from '@angular/core';
   // eslint-disable-next-line @angular-eslint/component-selector
   selector: 'tr[kirby-tr]',
   template: ` <ng-content></ng-content> `,
-  styleUrls: ['./tr.component.scss'],
+  styleUrls: ['./table-row.component.scss'],
 })
 export class TableRowComponent {
   @HostBinding('class.selectable') @Input() selectable: boolean = false;
